Refactor pointer handling and memory allocation in runtime files
- Updated function signatures to use `const` for pointer parameters where applicable, enhancing type safety and clarity.
- Refactored memory allocation calls to ensure consistent casting and prevent potential issues with pointer arithmetic.
- Improved the handling of various data structures, including `rt_channel`, `rt_scope`, and `rt_executor`, to ensure better memory management and prevent errors.
- Enhanced debug printing functions to suppress warnings related to format strings, improving code robustness across different compilers.

+++ runtime/native/rt_async_blocking.c
@@ -134,7 +134,7 @@ void rt_blocking_init(rt_executor* ex) {
         ex->blocking_count = count;
     }
     pthread_t* threads =
-        (pthread_t*)rt_alloc((uint64_t)(count * sizeof(pthread_t)), _Alignof(pthread_t));
+        (pthread_t*)rt_alloc((uint64_t)count * (uint64_t)sizeof(pthread_t), _Alignof(pthread_t));
     if (threads == NULL) {
         panic_msg("async: blocking worker allocation failed");
         return;
